% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidySpeciesData.R
\name{addTaxonId}
\alias{addTaxonId}
\title{Adds \code{internalTaxonId} OR \code{scientificName} column}
\usage{
addTaxonId(species_data)
}
\arguments{
\item{species_data}{a tibble of data downloaded from an IUCN Redlist search and
imported using \code{\link{importList}}}
}
\value{
a list with species
\code{internalTaxonId}; \code{scientificName}; and \code{redlistCriteria}
split into three levels.
}
\description{
\code{addTaxonId} adds \code{internalTaxonId} OR \code{scientificName} column
to any dataframe in species_data missing one of these columns by using
dplyr::right_join on each data frame in list and \code{internalTaxonId} and
\code{scientificName}) selected from either the 'species_assessments' or
'assessments' element
}
\examples{
tidyCriteria(species_data[['redlistCriteria']])
}
