% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCredits.R
\name{cleanFullName}
\alias{cleanFullName}
\title{cleans a vector of full names}
\usage{
cleanFullName(names)
}
\arguments{
\item{names}{a vector of names to be cleaned}
}
\value{
a vector of cleaned names
}
\description{
cleans a vector of full names
}
\section{Used in}{
 \code{\link{tidyCredits}}
}

\examples{
tidyCredits(species_data)
}
