% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBib.R
\name{createShortRef}
\alias{createShortRef}
\title{creates short ref for \code{bibliography} entries}
\usage{
createShortRef(bibliography)
}
\arguments{
\item{bibliography}{an intermediate object in \code{\link{tidyBib}} made up
of \code{species_data[['references']]} plus \code{referenceId}}
}
\value{
bibliography with added column \code{shortRef}
}
\description{
creates short ref for \code{bibliography} entries
}
\section{Used in}{
 \code{\link{journalName}}
}

\examples{
creatShortRef(bibliography)
}
