% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCriteria.R
\name{criteriaToColumns}
\alias{criteriaToColumns}
\title{Formats a string of criteria of the same level 1 to one criterion per row}
\usage{
criteriaToColumns(input, pattern, firstHalf, secondHalf)
}
\arguments{
\item{input}{a column of Redlist criteria strings seperated by pattern parameter.}

\item{pattern}{the character or string used to seperate Redlist criteria
within input, in regex format.}

\item{firstHalf}{the string common to all criteria in a string but not
explicitly stated after the first, in A2cd this is A or '^[A-Z]'}

\item{secondHalf}{the string unique to that criteria in a string, in A2cd
this is 2cd or '[0-9].$'}
}
\value{
tibble with ncol=max number of elements in string, each col named V1,V2 etc.
}
\description{
Formats a string of criteria of the same level 1 to one criterion per row
}
\section{Used in}{
 \code{\link{tidyCriteria}}
}

\examples{
criteriaToColumns(species_data[['redlistCriteria']]$redlistCriteria,'[+]','^.','[0-9].*$')
}
