% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCriteria.R
\name{criteriaToColumnsL2}
\alias{criteriaToColumnsL2}
\title{Formats a Redlist criterion into its sub-categories, dependency of tidyCriteria}
\usage{
criteriaToColumnsL2(input)
}
\arguments{
\item{input}{a column of seperated Redlist criteria e.g. A2cd}
}
\value{
tibble with ncol=max number of subcategories, each col named V1,V2 etc.
}
\description{
Formats a Redlist criterion into its sub-categories, dependency of tidyCriteria
}
\section{Used in}{
 \code{\link{tidyCriteria}}
}

\examples{
criteriaToColumnsL2(resultFromPivotToColumns$redlistCriteria)
}
