% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBib.R
\name{dateAccessed}
\alias{dateAccessed}
\title{Selects access date}
\usage{
dateAccessed(bibliography)
}
\arguments{
\item{bibliography}{an intermediate object in \code{\link{tidyBib}} made up
of \code{species_data[['references']]} plus \code{referenceId} and results of
\code{\link{pageNumbers}}}
}
\value{
tibble with \code{dayAccessed},\code{monthAccessed} and
\code{yearAccessed} if present
}
\description{
Selects access date
}
\section{Used in}{
 \code{\link{tidyCitation}}
}

\examples{
journal(bibliography)
}
