% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listManagement.R
\name{exportList}
\alias{exportList}
\title{Export list of dataframes to any format, requires \link[rio] for non-csv filetypes}
\usage{
exportList(list, directory, prefix, fileType)
}
\arguments{
\item{list}{the list to be exported}

\item{directory}{the directory the list is to be exported to}

\item{prefix}{the prefix that will be added to file names}

\item{fileType}{the intended export filetype e.g. 'csv'}
}
\description{
Export list of dataframes to any format, requires \link[rio] for non-csv filetypes
}
\examples{
exportList(speciesData,'/exported_files','species_data','csv')
}
