% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCredits.R
\name{firstNameToInitials}
\alias{firstNameToInitials}
\title{converts a vector of names into initials}
\usage{
firstNameToInitials(names)
}
\arguments{
\item{names}{a vector of names to be cleaned}
}
\value{
a value containing initials with no spaces or punctuation between
}
\description{
converts a vector of names into initials
}
\section{Used in}{
 \code{\link{cleanFullName}}
}

