% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBib.R
\name{inTextRefs}
\alias{inTextRefs}
\title{returns in text references}
\usage{
inTextRefs(text)
}
\arguments{
\item{text}{the text to be searched for full text references}
}
\value{
\code{refs} a vector of the full text references in \code{text}
}
\description{
\code{inTextRefs} searches for in text references using the regex pattern
'[(][^)]*[1-2][0-9][0-9][0-9][^(]*[)]'
}
\section{Used in}{
 \code{\link{}}
}

