% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidySpeciesData.R
\name{joinAll}
\alias{joinAll}
\title{Joins all elements in a list}
\usage{
joinAll(list)
}
\arguments{
\item{list}{the list to be joined, rows will be lost is all elements do not
have the same number of rows}
}
\value{
a dataframe with nrow = nrow(list[[1]]) and all columns from all
elements of \code{list}
}
\description{
\code{joinAll} uses dplyr::left_join() to merge all the elements in a list
of tibbles
}
