% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBib.R
\name{journalName}
\alias{journalName}
\title{Selects journal name}
\usage{
journalName(bibliography)
}
\arguments{
\item{bibliography}{an intermediate object in \code{\link{tidyBib}} made up
of \code{species_data[['references']]} plus \code{referenceId}}
}
\value{
tibble with \code{journalName} and \code{inSCImago}
}
\description{
Selects journal name
}
\section{Used in}{
 \code{\link{tidyCitation}}
}

\examples{
journal(bibliography)
}
