% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBib.R
\name{matchRefs}
\alias{matchRefs}
\title{matches in text references with references from bibliography}
\usage{
matchRefs(species_data, max_distance)
}
\arguments{
\item{species_data}{a tibble of data downloaded from an IUCN Redlist search
and imported using \code{\link{importList}}}

\item{max_distance}{the maximum accespted standard levenschtien distance
between an in text reference and its match}
}
\value{
\code{species_data} with two extra columns in \code{in_text}:
\code{bestMatch} and \code{referenceId}
}
\description{

}
\section{Used in}{
 \code{\link{}}
}

