% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBib.R
\name{pageNumbers}
\alias{pageNumbers}
\title{Selects page numbers from citation if listed at end}
\usage{
pageNumbers(bibliography)
}
\arguments{
\item{bibliography}{an intermediate object in \code{\link{tidyBib}} made up
of \code{species_data[['references']]} plus reference ID}
}
\value{
tibble with \code{pages},\code{startPage} and \code{endPage}
}
\description{
Selects page numbers from citation if listed at end
}
\section{Used in}{
 \code{\link{tidyCitation}}
}

\examples{
pageNumbers(bibliography)
}
