% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCriteria.R
\name{pasteFirstToAll}
\alias{pasteFirstToAll}
\title{Pastes the value in the first column of each row in front of each value in that row}
\usage{
pasteFirstToAll(df)
}
\arguments{
\item{df}{a tibble with value to be pasted in the first column}
}
\value{
a tibble with the value in the last column of each row in front of
each value in that row and first column removed
}
\description{
Pastes the value in the first column of each row in front of each value in that row
}
\section{Used in}{
 \code{\link{criteriaToColumnsL2}}
}

