% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCriteria.R
\name{pasteLastToAll}
\alias{pasteLastToAll}
\title{Pastes the value in the last column of each row in front of each value in that row}
\usage{
pasteLastToAll(df)
}
\arguments{
\item{df}{a tibble with value to be pasted in the last column}
}
\value{
a tibble with the value in the last column of each row in front of
each value in that row and last columns removed
}
\description{
Pastes the value in the last column of each row in front of each value in that row
}
\section{Used in}{
 \code{\link{criteriaToColumns}}
}

