% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBib.R
\name{refDistances}
\alias{refDistances}
\title{fuzzy match references
#' @param species_data a tibble of data downloaded from an IUCN Redlist search
and imported using \code{\link{importList}}}
\usage{
refDistances(species_data)
}
\value{
tibble with \code{internaltaxonId}, \code{xPosition}, \code{xName},
\code{yPosition}, \code{yName} and \code{adist}
}
\description{
fuzzy match references
#' @param species_data a tibble of data downloaded from an IUCN Redlist search
and imported using \code{\link{importList}}
}
\section{Used in}{
 \code{\link{matchRefs}}
}

