% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidySpeciesData.R
\name{removeNA}
\alias{removeNA}
\title{Removes rows containing \code{NA} in all but two columns}
\usage{
removeNA(species_data)
}
\arguments{
\item{species_data}{a tibble of data downloaded from an IUCN Redlist search
and imported using \code{\link{importList}}}
}
\description{
Removes rows containing \code{NA} in all but two columns
}
\section{Used in}{
 \code{\link{tidySpeciesData}}
}

\examples{
removeNA(species_data)
}
