% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBib.R
\name{SCImagoJournal}
\alias{SCImagoJournal}
\title{Searches for the journal names in \code{rankings}}
\usage{
SCImagoJournal(bibliography)
}
\arguments{
\item{bibliography}{an intermediate object in \code{\link{tidyBib}} made up
of \code{species_data[['references']]} plus \code{referenceId}}
}
\value{
tibble with \code{journalName} and \code{inSCImago}
}
\description{
Note: this takes a while as there are many, many journals, and each has an abbreviation!
}
\section{Used in}{
 \code{\link{journalName}}
}

\examples{
journal(bibliography)
}
