% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBib.R
\name{speciesInTextRefs}
\alias{speciesInTextRefs}
\title{Finds the in text references for each text in each species' redlist profile}
\usage{
speciesInTextRefs(species_data)
}
\arguments{
\item{species_data}{a tibble of data downloaded from an IUCN Redlist search
and imported using \code{\link{importList}}}
}
\value{
\code{species_data} with an extra tibble \code{in_text}
}
\description{
Finds the in text references for each text in each species' redlist profile
}
\section{Used in}{
 \code{\link{}}
}

