% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTo.R
\name{splitToColumns}
\alias{splitToColumns}
\title{Splits concatenated observations into multiple columns}
\usage{
splitToColumns(x, pattern)
}
\arguments{
\item{x}{the vector or column of concatenated observations to be split}

\item{pattern}{a regex representation of the characters used to split elements
in the concatenated column}
}
\value{
a tibble with one row per concatenated observation, and a column for
each observation within a row
}
\description{
\code{splitToColumns} will take a vector/ column of concatenated observations
and split them into one observation per column, maintaining the rows they were
in before
}
\section{Used in}{
 \code{\link{splitToTidy}}
}

\examples{
splitToColumns(species_data[['assessments']]$realms, '[|]')
}
