% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTo.R
\name{splitToTidy}
\alias{splitToTidy}
\title{Splits concatenated observations into multiple columns}
\usage{
splitToTidy(columns, pattern, IDs, name)
}
\arguments{
\item{columns}{the columns containing concatenated observations to be split}

\item{pattern}{a regex representation of the characters used to split elements
in the concatenated column}

\item{IDs}{the identifier for each row in x, e.g. internalTaxonId}

\item{name}{the names to be given to each tibble}
}
\value{
a list with a tibble containing IDs and one result per row for each
column in \cpde{columns}
}
\description{
\code{splitToTidy} # will take (e.g. Palearctic|Afrotropical) a tibble of
concatenated columns and species IDs and will return a list with tibbles
with IDs and split results for each column
}
\examples{
splitToTidy(c(species_data[['assessments']]$realms, species_data[['assessments']]$system),
c('[|]','[|]'), species_data[['assessments']]$internalTaxonId,c('realms','system'))
}
