% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidySpeciesData.R
\name{taxonIdToChar}
\alias{taxonIdToChar}
\title{Converts all instances of \code{internalTaxonId} to \code{\link{as.character}}}
\usage{
taxonIdToChar(species_data)
}
\arguments{
\item{species_data}{a tibble of data downloaded from an IUCN Redlist search
and imported using \code{\link{importList}}}
}
\description{
Converts all instances of \code{internalTaxonId} to \code{\link{as.character}}
}
\section{Used in}{
 \code{\link{tidySpeciesData}}
}

\examples{
taxonIdToChar(species_data)
}
