% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidySpeciesData.R
\name{tidyAssessments}
\alias{tidyAssessments}
\title{Combines all line-per-species tibbles}
\usage{
tidyAssessments(species_data)
}
\arguments{
\item{species_data}{a tibble of data downloaded from an IUCN Redlist search
and imported using \code{\link{importList}}}
}
\value{
\code{species_data} with all line-per-species tibbles combined into
\code{assessments}
}
\description{
Combines all line-per-species tibbles
}
\examples{
tidyAssessments(species_data)
#' @section Used in: \code{\link{tidySpeciesData}}
}
