% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBib.R
\name{tidyAuthors}
\alias{tidyAuthors}
\title{Tidies authors of redlist bibliographies into a tidy dataframe}
\usage{
tidyAuthors(bibliography)
}
\arguments{
\item{bibliography}{an intermediate object in \code{\link{tidyBib}} made up
of \code{species_data[['references']]} plus reference ID}
}
\value{
tibble with \code{internalTaxonId}, \code{referenceId} and
\code{author} (cleaned author name)
}
\description{
Tidies authors of redlist bibliographies into a tidy dataframe
}
\section{Used in}{
 \code{\link{tidyBib}}
}

\examples{
tidyAuthors(bibliography)
}
