% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBib.R
\name{tidyBib}
\alias{tidyBib}
\title{Tidies redlist bibliographies into a tidy dataframe}
\usage{
tidyBib(species_data)
}
\arguments{
\item{species_data}{a tibble of data downloaded from an IUCN Redlist search
and imported using \code{\link{importList}}}
}
\value{
\code{species_data} with an extra tibble for bibliography authors and
tidied bibliography entries
}
\description{
Tidies redlist bibliographies into a tidy dataframe
}
\section{Used in}{
 \code{\link{tidySpeciesData}}
}

\examples{
tidyBib(species_data)
}
