% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyConcat.R
\name{tidyCodes}
\alias{tidyCodes}
\title{Splits all numeric multi-level codes into columns}
\usage{
tidyCodes(species_data)
}
\arguments{
\item{species_data}{a tibble of data downloaded from an IUCN Redlist search
and imported using \code{\link{importList}}}
}
\value{
\code{species_data} with columns added splitting codes into one level
per column
}
\description{
Splits all numeric multi-level codes into columns
}
\section{Used in}{
 \code{\link{tidyConcat}}
}

\examples{
tidyCodes(species_data)
}
