% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyConcat.R
\name{tidyConcat}
\alias{tidyConcat}
\title{Splits all columns with concatenated observations into tidy tibbles}
\usage{
tidyConcat(species_data)
}
\arguments{
\item{species_data}{a tibble of data downloaded from an IUCN Redlist search
and imported using \code{\link{importList}}}
}
\value{
\code{species_data} with tidy tibbles for systems, realms and
redlistCriteria and removes these columns from assessments
}
\description{
Splits all columns with concatenated observations into tidy tibbles
}
\section{Used in}{
 \code{\link{tidySpeciesData}}
}

\examples{
tidyConcat(species_data)
}
