% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCredits.R
\name{tidyCredits}
\alias{tidyCredits}
\title{Formats \code{credits} into tidy data}
\usage{
tidyCredits(species_data)
}
\arguments{
\item{species_data}{a tibble of data downloaded from an IUCN Redlist search and
imported using \code{\link{importList}}}
}
\value{
a tibble with \code{internalTaxonId}, \code{scientificName},
\code{type}, \code{name} and \code{organisation}.
}
\description{
Formats \code{credits} into tidy data
}
\section{Used in}{
 \code{\link{tidyConcat}}
}

\examples{
tidyCredits(species_data)
}
