% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCriteria.R
\name{tidyCriteria}
\alias{tidyCriteria}
\title{Formats \code{redlistCriteria} into tidy data}
\usage{
tidyCriteria(redCriteria)
}
\arguments{
\item{redCriteria}{a tibble with \code{redlistCriteria}, \code{internalTaxonId}
and \code{scientificName}.}
}
\value{
Tibble with each species-criterion as a row and columns
\code{internalTaxonId}; \code{scientificName}; and \code{redlistCriteria}
split into three levels.
}
\description{
Formats \code{redlistCriteria} into tidy data
}
\section{Used in}{
 \code{\link{tidySpeciesData}}
}

\examples{
tidyCriteria(species_data[['redlistCriteria']])
}
