% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyNames.R
\name{tidyNames}
\alias{tidyNames}
\title{Formats \code{common_names} and \code{synonyms} into a tidy tibble}
\usage{
tidyNames(species_data)
}
\arguments{
\item{species_data}{a list of data downloaded from the IUCN Redlist and
imported using \code{\link{importList}}}
}
\value{
Tibble with each name as a row and columns \code{internalTaxonId},
\code{name},\code{language},\code{source}
}
\description{
Formats \code{common_names} and \code{synonyms} into a tidy tibble
}
\section{Used in}{
 \code{\link{tidySpeciesData}}
}

\examples{
tidyNames(species_data)
}
