% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidySpeciesData.R
\name{tidySpeciesData}
\alias{tidySpeciesData}
\title{Formats Redlist downloads into tidy data}
\usage{
tidySpeciesData(species_data)
}
\arguments{
\item{species_data}{a tibble of data downloaded from an IUCN Redlist search and
imported using \code{\link{importList}}}
}
\value{
a list with species
}
\description{
Formats Redlist downloads into tidy data
}
\examples{
tidyCriteria(species_data[['redlistCriteria']])
}
