% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyNames.R
\name{tidySynonyms}
\alias{tidySynonyms}
\title{Formats \code{synonyms} into a tidy tibble}
\usage{
tidySynonyms(synonyms)
}
\arguments{
\item{species_data}{a list of data downloaded from the IUCN Redlist and
imported using \code{\link{importList}}}
}
\value{
Tibble with each alternate scientific name as a row and columns
\code{scientificName} and \code{name}
}
\description{
Formats \code{synonyms} into a tidy tibble
}
\section{Used in}{
 \code{\link{tidyNames}}
}

\examples{
tidySynonyms(species_data)
}
