% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBib.R
\name{volumeIssue}
\alias{volumeIssue}
\title{Selects volume and issue}
\usage{
volumeIssue(bibliography)
}
\arguments{
\item{bibliography}{an intermediate object in \code{\link{tidyBib}} made up
of \code{species_data[['references']]} plus \code{referenceId}}
}
\value{
tibble with \code{volume} and \code{issue} if present
}
\description{
Selects volume and issue
}
\section{Used in}{
 \code{\link{tidyCitation}}
}

\examples{
journal(bibliography)
}
