% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCriteria.R
\name{tidyCriteria}
\alias{tidyCriteria}
\title{Formats \code{redlistCriteria} into tidy data}
\usage{
tidyCriteria(species_data)
}
\arguments{
\item{species_data}{a tibble of data downloaded from an IUCN Redlist search
and imported using \code{\link{importList}}}
}
\value{
Tibble with each species-criterion as a row and columns
\code{internalTaxonId}; \code{scientificName}; and \code{redlistCriteria}
split into three levels.
}
\description{
Formats \code{redlistCriteria} into tidy data
}
\section{Used in}{
 \code{\link{tidyConcat}}
}

\examples{
tidyCriteria(species_data[['redlistCriteria']])
}
